fix(manageTasks): Made the required changes to the Query and Command Process, implemented the CRUD required.

+++ AidManager.API/ManageTasks/Application/Internal/CommandServices/TaskCommandService.cs
@@ -7,7 +7,7 @@ using AidManager.API.Shared.Domain.Repositories;
 
 namespace AidManager.API.ManageTasks.Application.Internal.CommandServices;
 
-public class TaskCommandService(ITaskRepository taskRepository, IUnitOfWork unitOfWork, IProjectRepository projectRepository, ExternalUserService externalUserService) : ITaskCommandService
+public class TaskCommandService(ITaskRepository taskRepository, IUnitOfWork unitOfWork, IProjectRepository projectRepository, ExternalUserService externalUserService, IProjectCommandService projectCommandService) : ITaskCommandService
 {
     public async Task<TaskItem> Handle(CreateTaskCommand command)
     {
@@ -19,11 +19,12 @@ public class TaskCommandService(ITaskRepository taskRepository, IUnitOfWork unit
             throw new Exception($"Project with id {command.ProjectId} does not exist.");
         }
         
-        var username = await externalUserService.GetUserFullnameById(command.AssigneeId);
+        var user = await externalUserService.GetUserById(command.AssigneeId);
+        var fullname = user.FirstName + " " + user.LastName;
         
+        var task = new TaskItem(command, fullname);
         
-        var task = new TaskItem(command, username);
-        
+        await projectCommandService.Handle(new AddTeamMemberCommand(command.AssigneeId, command.ProjectId));
         
         try
         {
@@ -41,27 +42,42 @@ public class TaskCommandService(ITaskRepository taskRepository, IUnitOfWork unit
     
     public async Task<TaskItem> Handle(UpdateTaskCommand command)
     {
-        var username = await externalUserService.GetUserFullnameById(command.AssigneeId);
-        var task = await taskRepository.FindByIdAsync(command.Id);
-        Console.WriteLine("command service called");
+        
+        bool exists = await projectRepository.ExistsProject(command.ProjectId);
+
+        if (!exists)
+        {
+            throw new Exception($"Project with id {command.ProjectId} does not exist.");
+        }
+        
+        var username = await externalUserService.GetUserById(command.AssigneeId);
+        var task = await taskRepository.GetTaskById(command.Id);
+        var fullname = username.FirstName + " " + username.LastName;
+        
+
         if (task is null) throw new Exception("Task not found");
-        task.UpdateTask(command, username);
-        Console.WriteLine("task Aggregate updated");
+        
+        task.UpdateTask(command, fullname);
         await taskRepository.Update(task);
-        Console.WriteLine("task updated in repository");
         await unitOfWork.CompleteAsync();
-        Console.WriteLine("unit of work completed");
         return task;
     }
     
     public async Task<TaskItem> Handle(DeleteTaskCommand command)
     {
-        var task = await taskRepository.FindByIdAsync(command.Id);
+        
+        bool exists = await projectRepository.ExistsProject(command.ProjectId);
+
+        if (!exists)
+        {
+            throw new Exception($"Project with id {command.ProjectId} does not exist.");
+        }
+        
+        var task = await taskRepository.GetTaskById(command.Id);
         if (task is null) throw new Exception("Task not found");
         await taskRepository.Remove(task);
         await unitOfWork.CompleteAsync();
         return task;
     }
-    
     
 }
